Split: apertium_tagger.cc to tagger.h & tagger.cc
