+++ apertium/tagger.h
@@ -13,19 +13,50 @@
 // You should have received a copy of the GNU General Public License
 // along with this program; if not, see <http://www.gnu.org/licenses/>.
 
-#ifndef APERTIUM_TAGGER_H
-#define APERTIUM_TAGGER_H
+#ifndef TAGGER_H
+#define TAGGER_H
 
 #include "apertium_config.h"
 
+#include "align.h"
+#include "basic_exception_type.h"
 #include "basic_stream_tagger.h"
 #include "basic_stream_tagger_trainer.h"
 #include "basic_tagger.h"
 #include "constructor_eq_delete.h"
+#include "err_exception.h"
+#include "exception.h"
 #include "file_tagger.h"
+#include "linebreak.h"
 #include "optional.h"
+#include "stream_5_3_1_tagger.h"
+#include "stream_5_3_1_tagger_trainer.h"
+#include "stream_5_3_2_tagger.h"
+#include "stream_5_3_2_tagger_trainer.h"
+#include "stream_5_3_3_tagger.h"
+#include "stream_5_3_3_tagger_trainer.h"
+#include <apertium/perceptron_tagger.h>
+#include <apertium/file_morpho_stream.h>
+#include <apertium/hmm.h>
+#include <apertium/lswpost.h>
+#include <apertium/tagger_word.h>
+#include <apertium/shell_utils.h>
+
+#include <lttoolbox/lt_locale.h>
 
 #include "getopt_long.h"
+#include <cerrno>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+#include <fstream>
+#include <iomanip>
+#include <ios>
+#include <iostream>
+#include <locale>
+#include <sstream>
+#include <string>
+#include <unistd.h>
 #include <string>
 
 namespace Apertium {
@@ -94,4 +125,4 @@ private:
 };
 }
 
-#endif // APERTIUM_TAGGER_H
+#endif // TAGGER_H
